Enable reuse of BackOfficeSecurityRequirementsOperationFilter for custom APIs
